Бот/salesbot_final(копия)/salesbot_INSTALLER_DELTA_v3_FULL/_installer/runtime/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv>=1.0.0
requests>=2.31.0
